gen-ai-playground/backend/requirements.txt
fastapi[standard]
python-dotenv
requests
pymongo
bcrypt
PyJWT
pytest
pytest-mock
mongomock
httpx
verda